Replace geocoding Lua scripts with Redis transactions
Keep optimistic acknowledgement scoped to a single user and day, deferring conflicting updates without retry loops. Move bounded refresh orchestration into a service and leave the stats job as its entry point.
